[level 4] Title: 무지의 먹방 라이브, Time: 113.37 ms, Memory: 65.5 MB -BaekjoonHub

+++ 프로그래머스/4/42891. 무지의 먹방 라이브/무지의 먹방 라이브.java
@@ -7,8 +7,8 @@ class Solution {
         
         Arrays.sort(sortFood);
         
-        long roof = 0;
-        long roofCount = 0;
+        int roof = 0;
+        int roofCount = 0;
         long count = 0;
         for (int i = 0; i < sortFood.length; i++) {
             for (int j = 0; j < sortFood[i] - roof; j++) {
